+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/system/ScoringDBMapping.java
@@ -2,7 +2,16 @@ package de.mephisto.vpin.restclient.system;
 
 public class ScoringDBMapping {
   private String rom;
+  private String tableName;
   private String textFile;
+
+  public String getTableName() {
+    return tableName;
+  }
+
+  public void setTableName(String tableName) {
+    this.tableName = tableName;
+  }
 
   public String getRom() {
     return rom;