+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/roms/ScanResult.java
@@ -43,6 +43,9 @@ public class ScanResult {
   }
 
   public void setHsFileName(String hsFileName) {
+    while (hsFileName.contains("\\")) {
+      hsFileName = hsFileName.substring(hsFileName.indexOf("\\") + 1);
+    }
     this.hsFileName = hsFileName;
   }