+++ RELEASE_NOTES.md
@@ -34,6 +34,7 @@ Even if you table scan resolved the correct ROM name of a table.
 - Added checkbox to enable fuzzy lookup logic for the backglass server in the "Backglass Server" preferences.
 - Improved navigation between the Table Data Manager and the Table Asset Manager, including previous and next button.
 - Added "Invalidate All" as split-button to the "Invalidate" button in the overview toolbar.
+- Added auto-full rules for the VPX meta-data fields notes and rules, so that these values are written to the "notes" and "details" fields of PinUP Popper 1.5. 
 - VPS update flags are now resetted for the following uploads: backglass, table replacements, altsound, altcolor, puppack and pov files/archives.
 - The automatic writing from mapped VPS data to PinUP popper has been removed again for: initial scans, table replacements and clones.
 - The "Table Import" dialog now provides a progress bar, so that the user gets feedback about the progress when a bunch of tables is installed at once.